+++ magpylib/_lib/getBvector.py
@@ -25,13 +25,13 @@ from magpylib._lib.fields.PM_Box_vector import Bfield_BoxV
 from magpylib._lib.fields.PM_Cylinder_vector import Bfield_CylinderV
 from magpylib._lib.fields.PM_Sphere_vector import Bfield_SphereV
 from magpylib._lib.fields.Moment_Dipole_vector import Bfield_DipoleV
+from magpylib._lib.fields.Current_CircularLoop_vector import Bfield_CircularCurrentLoopV
 from magpylib._lib.mathLib_vector import QconjV, QrotationV, QmultV, getRotQuatV, angleAxisRotationV_priv
 import numpy as np
 
-def getBv_magnet(type,MAG,DIM,POSo,POSm,ANG=[],AX=[],ANCH=[],Nphi0=50):
+def getBv_magnet(type,MAG,DIM,POSm,POSo,ANG=[],AX=[],ANCH=[],Nphi0=50):
     """
     Calculate the field of magnets using vectorized performance code.
-    This function is faster than getB when more than ~5 fields are evaluated.
 
     Parameters
     ----------
@@ -106,16 +106,77 @@ def getBv_magnet(type,MAG,DIM,POSo,POSm,ANG=[],AX=[],ANCH=[],Nphi0=50):
 # -------------------------------------------------------------------------------
 
 
-def getBv_current():
-    return 0
+def getBv_current(type,CURR,DIM,POSm,POSo,ANG=[],AX=[],ANCH=[]):
+    """
+    Calculate the field of currents using vectorized performance code.
+
+    Parameters
+    ----------
+
+    type : string
+        source type either 'circular' or 'line'
+
+    MAG : Nx3 numpy array float [mT]
+        vector of N magnetizations.
+
+    DIM : NxY numpy array float [mm]
+        vector of N dimensions for each evaluation. The form of this vector depends
+        on the source type. Y=3/2/1 for box/cylinder/sphere
+
+    POSo : Nx3 numpy array float [mm]
+        vector of N positions of the observer.
+    
+    POSm : Nx3 numpy array float [mm]
+        vector of N initial source positions. These positions will be adjusted by
+        the given rotation parameters.
+
+    ANG=[] : length M list of size N numpy arrays float [deg]
+       Angles of M subsequent rotation operations applied to the N-sized POSm and
+       the implicit source orientation.
+    
+    AX=[] : length M list of Nx3 numpy arrays float []
+        Axis vectors of M subsequent rotation operations applied to the N-sized
+        POSm and the implicit source orientation.
+    
+    ANCH=[] : length M list of Nx3 numpy arrays float [mm]
+        Anchor positions of M subsequent rotations applied ot the N-sized POSm and
+        the implicit source orientation.
+    """
+    N = len(POSo)
+
+    Q = np.zeros([N,4])
+    Q[:,0] = 1              # init orientation
+    
+    Pm = POSm               #initial position
+
+    #apply rotation operations
+    for ANGLE,AXIS,ANCHOR in zip(ANG,AX,ANCH):
+        Q = QmultV(getRotQuatV(ANGLE,AXIS),Q)
+        Pm = angleAxisRotationV_priv(ANGLE,AXIS,Pm-ANCHOR)+ANCHOR
+
+    # transform into CS of source
+    POSrel = POSo-Pm        #relative position
+    Qc = QconjV(Q)          #orientierung
+    POSrot = QrotationV(Qc,POSrel)  #rotation der pos in das CS der Quelle
+
+    # calculate field
+    if type == 'circular':
+        Brot = Bfield_CircularCurrentLoopV(CURR, DIM, POSrot)
+    else:
+        print('Bad type')
+        return 0
+    
+    # transform back
+    B = QrotationV(Q,Brot)  #rückrotation des feldes
+    
+    return B
 
 # -------------------------------------------------------------------------------
 # -------------------------------------------------------------------------------
 
-def getBv_moment(type,MOM,POSo,POSm,ANG=[],AX=[],ANCH=[]):
+def getBv_moment(type,MOM,POSm,POSo,ANG=[],AX=[],ANCH=[]):
     """
     Calculate the field of magnetic moments using vectorized performance code.
-    This function is faster than getB when more than ~5 fields are evaluated.
 
     Parameters
     ----------