+++ tests/test_MathVector.py
@@ -4,6 +4,8 @@ from numpy import array, amax
 from magpylib._lib.mathLib import getRotQuat, axisFromAngles, anglesFromAxis, angleAxisRotation, elliptic
 from magpylib.math import randomAxisV
 import numpy as np
+from magpylib._lib.mathLib import ellipticK, ellipticE, ellipticPi
+from magpylib._lib.mathLib_vector import ellipticKV, ellipticEV, ellipticPiV
 
 
 def test_QV():
@@ -106,3 +108,21 @@ def test_ellipticV():
     solV = ellipticV(INP[:,0],INP[:,1],INP[:,2],INP[:,3])
 
     assert np.amax(abs(solC-solV)) < 1e-10
+
+
+
+def test_ellipticSpecialCases():
+    X = np.linspace(0,.9999,33)
+    Y = np.linspace(0,.9999,33)[::-1]
+
+    solV = ellipticKV(X)
+    solC = np.array([ellipticK(x) for x in X])
+    assert np.amax(abs(solV-solC)) < 1e-15
+
+    solV = ellipticEV(X)
+    solC = np.array([ellipticE(x) for x in X])
+    assert np.amax(abs(solV-solC)) < 1e-15
+
+    solV = ellipticPiV(X,Y)
+    solC = np.array([ellipticPi(x,y) for x,y in zip(X,Y)])
+    assert np.amax(abs(solV-solC)) < 1e-15